Device details

+++ hardware/src/main/java/org/xapik/hardware/device/device/main/DeviceRepository.java
@@ -3,6 +3,6 @@ package org.xapik.hardware.device.device.main;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.xapik.hardware.device.device.main.model.DeviceEntity;
 
-public interface DeviceRepository extends JpaRepository<DeviceEntity, Integer> {
+public interface DeviceRepository extends JpaRepository<DeviceEntity, Long> {
 
 }
